SUMMARY:   Fixed a simple unnoticed bug where MPI_Reduce was passed an MPI_INT when it should have been an MPI_DOUBLE.  This effects anyone using ML_gsum_vec_double().
     BUGS:      [???]

     CHECKLIST:

     1)   Did you add any source files?  [???]
     1.a) Did you do a 'cvs add'?  [???]
     1.b) Did you add them to a Makefile.am file in some way?  [???]
     1.c) Did you do a new bootstrap?  [???]

     2)   Did you update or change the build system?  [???]
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  [???]
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  [???]

     3) Did you rerun the test suite on at least the packages that you
        changed?  [???]

     COMMENTS:

     [???]

(checklist version 2007-06-20)

+++ packages/ml/src/Utils/ml_utils.c
@@ -1101,7 +1101,7 @@ void ML_gsum_vec_double(double **tvals, double **tvals2, int length, ML_Comm *co
 #ifdef ML_MPI
   double *tmpptr;
 
-  MPI_Allreduce((void *) *tvals,(void *) *tvals2, length, MPI_INT, MPI_SUM,
+  MPI_Allreduce((void *) *tvals,(void *) *tvals2, length, MPI_DOUBLE, MPI_SUM,
                 comm->USR_comm);
   tmpptr = *tvals;
   *tvals = *tvals2;
